Update ftp_client
here the client transfers the file to the server. It prompts  for the path of the file to be transmitted. once entered , if this client is present in the server acceptance list , it prompts if we wish to transmit the file. once transmitted it calculates the time taken for the file to get transmitted.